Cardano.Ledger.Spec.STS.UTXO

Description

UTXO transition system

Documentation

data UTXO

data UTxOEnv

Constructors

UTxOEnv UTxO PParams

data UTxOState

Constructors

UTxOState UTxO Lovelace

data family PredicateFailure a ∷ Type

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

utxo ∷ UTxOState → UTxO

utxo0 ∷ UTxOEnv → UTxO

pps ∷ UTxOEnv → PParams

reserves ∷ UTxOState → Lovelace