Ledger.Core.Omniscient

Description

Functions that are intended to be used in tests, but not in the executable specifications. These functions are not possible in practice (like recovering the signed data from the signature only), but are possible for the abstract data we use in the specifications.

Documentation

signatureVKey ∷ Sig a → VKey

Extract the verifying key of a signature. This is useful when elaborating an abstract signature into a concrete one.

signatureData ∷ Sig a → a

Extract the signature data.

skey ∷ VKey → SKey

Get the signing key from the verification key. We use this in the generators, where we need to generate signed data for a given verification key (e.g. one that appears in the delegation map) for which we don't have the signing key.

signWithGenesisKey ∷ VKeyGenesis → a → Sig a

Sign using a genesis verifying key.

See skey for details about situations where this function is used.