Ledger.Update

Documentation

pattern DoesNotVerify ∷ PredicateFailure UPREG

pattern AVSigDoesNotVerify ∷ PredicateFailure ADDVOTE

pattern AlreadyProposedPv ∷ PredicateFailure UPPVV

pattern AlreadyProposedSv ∷ PredicateFailure UPSVV

pattern CannotFollowPv ∷ PredicateFailure UPPVV

pattern CannotFollowSv ∷ PredicateFailure UPSVV

pattern CannotUpdatePv ∷ [UpdateConstraintViolation] → PredicateFailure UPPVV

pattern InvalidApplicationName ∷ PredicateFailure UPSVV

pattern InvalidSystemTags ∷ PredicateFailure UPSVV

pattern NoUpdateProposal ∷ UpId → PredicateFailure ADDVOTE

pattern NotGenesisDelegate ∷ PredicateFailure UPREG

pattern UPPVVFailure ∷ PredicateFailure UPPVV → PredicateFailure UPV

pattern UPSVVFailure ∷ PredicateFailure UPSVV → PredicateFailure UPV

pattern AVChangedInPVUpdate ∷ ApName → ApVer → Maybe (ApVer, Slot, Metadata) → PredicateFailure UPV

pattern ParamsChangedInSVUpdate ∷ PredicateFailure UPV

pattern PVChangedInSVUpdate ∷ PredicateFailure UPV

pattern UPVFailure ∷ PredicateFailure UPV → PredicateFailure UPREG

pattern VoteByNonGenesisDelegate ∷ VKey → PredicateFailure ADDVOTE

pattern RepeatVoteByGenesisDelegate ∷ VKey → PredicateFailure ADDVOTE

pattern ADDVOTEFailure ∷ PredicateFailure ADDVOTE → PredicateFailure UPVOTE

pattern S_HigherThanThdAndNotAlreadyConfirmed ∷ PredicateFailure UPVOTE

pattern S_CfmThdNotReached ∷ PredicateFailure UPVOTE

pattern S_AlreadyConfirmed ∷ PredicateFailure UPVOTE

pattern ProtVerUnknown ∷ ProtVer → PredicateFailure UPEND

pattern S_TryNextRule ∷ PredicateFailure UPEND

pattern CanAdopt ∷ ProtVer → PredicateFailure UPEND

pattern CannotAdopt ∷ ProtVer → PredicateFailure UPEND

pattern NotADelegate ∷ VKey → PredicateFailure UPEND

pattern UnconfirmedProposal ∷ UpId → PredicateFailure UPEND

pattern UPREGFailure ∷ PredicateFailure UPREG → PredicateFailure UPIREG

pattern UPVOTEFailure ∷ PredicateFailure UPVOTE → PredicateFailure UPIVOTE

pattern UpivoteFailure ∷ PredicateFailure UPIVOTE → PredicateFailure APPLYVOTES

pattern ApplyVotesFailure ∷ PredicateFailure APPLYVOTES → PredicateFailure UPIVOTES

pattern UPENDFailure ∷ PredicateFailure UPEND → PredicateFailure UPIEND

pattern NoPVBUMPFailure ∷ PredicateFailure PVBUMP

pattern PVBUMPFailure ∷ PredicateFailure PVBUMP → PredicateFailure UPIEC

data PParams

Protocol parameters.

Constructors

PParams

Fields

_maxBkSz ∷ !Natural

Maximum (abstract) block size in words

_maxHdrSz ∷ !Natural

Maximum (abstract) block header size in words

_maxTxSz ∷ !Natural

Maximum (abstract) transaction size in words

_maxPropSz ∷ !Natural

Maximum (abstract) update proposal size in words

_bkSgnCntT ∷ !BkSgnCntT

Fraction [0, 1] of the blocks that can be signed by any given key in a window of lenght _bkSgnCntW. This value will be typically between 1/5 and 1/4

_bkSlotsPerEpoch ∷ !SlotCount

Number of slots in an epoch. TODO: this should be removed since the number of slots per epoch should remain constant.

_upTtl ∷ !SlotCount

Update proposal TTL in slots

_scriptVersion ∷ !Natural

Script version

_upAdptThd ∷ !UpAdptThd

Update adoption threshold: a proportion of block issuers that have to endorse a given version to become candidate for adoption

_factorA ∷ !FactorA

Minimum fees per transaction

_factorB ∷ !FactorB

Additional fees per transaction size

newtype BkSgnCntT

Constructors

BkSgnCntT Double

newtype UpAdptThd

Constructors

UpAdptThd Double

newtype FactorB

Constructors

FactorB Int

newtype FactorA

Constructors

FactorA Int

data ProtVer

Protocol version

Constructors

ProtVer

Fields

_pvMaj ∷ Natural

_pvMin ∷ Natural

_pvAlt ∷ Natural

newtype UpId

Constructors

UpId Int

bkSgnCntT ∷ Lens' PParams BkSgnCntT

bkSlotsPerEpoch ∷ Lens' PParams SlotCount

factorA ∷ Lens' PParams FactorA

factorB ∷ Lens' PParams FactorB

maxBkSz ∷ Lens' PParams Natural

maxHdrSz ∷ Lens' PParams Natural

maxPropSz ∷ Lens' PParams Natural

maxTxSz ∷ Lens' PParams Natural

scriptVersion ∷ Lens' PParams Natural

upAdptThd ∷ Lens' PParams UpAdptThd

upTtl ∷ Lens' PParams SlotCount

data SwVer

Constructors

SwVer

Fields

_svName ∷ ApName

_svVer ∷ ApVer

newtype ApVer

Application version

Constructors

ApVer Natural

newtype ApName

Constructors

ApName String

pvAlt ∷ Lens' ProtVer Natural

pvMaj ∷ Lens' ProtVer Natural

pvMin ∷ Lens' ProtVer Natural

data UProp

Update proposal

Constructors

UProp

Fields

_upId ∷ UpId

_upIssuer ∷ VKey

_upParams ∷ PParams

_upPV ∷ ProtVer

_upSwVer ∷ SwVer

_upSig ∷ Sig UpSD

_upSTags ∷ Set STag

System tags involved in the update proposal.

_upMdt ∷ Metadata

Metadata required for performing software updates.

data Metadata

For now we do not have any requirements on metadata.

Constructors

Metadata

type STag = String

System tag, this represents a target operating system for the update (e.g. linux, win64, or mac32).

type UpSD = (ProtVer, PParams, SwVer, Set STag, Metadata)

Part of the update proposal which must be signed

svName ∷ Lens' SwVer ApName

svVer ∷ Lens' SwVer ApVer

data Vote

Constructors

Vote

Fields

_vCaster ∷ VKey

_vPropId ∷ UpId

_vSig ∷ Sig UpId

data UPREG

data UPV

Update proposal validity

data UPPVV

data UPSVV

Update Proposal Software Version Validation

data UpdateConstraintViolation

Violations on the constraints of the allowed values for new protocol parameters.

Constructors

BlockSizeTooLarge Natural (Threshold Natural)

TransactionSizeTooLarge Natural (Threshold Natural)

ScriptVersionTooLarge Natural (Threshold Natural)

ScriptVersionTooSmall Natural (Threshold Natural)

upId ∷ Lens' UProp UpId

upIssuer ∷ Lens' UProp VKey

upMdt ∷ Lens' UProp Metadata

upPV ∷ Lens' UProp ProtVer

upParams ∷ Lens' UProp PParams

upSTags ∷ Lens' UProp (Set STag)

upSig ∷ Lens' UProp (Sig UpSD)

upSwVer ∷ Lens' UProp SwVer

upSigData ∷ Lens' UProp UpSD

getUpSigData ∷ UProp → UpSD

mkUProp ∷ UpId → VKey → ProtVer → PParams → SwVer → Set STag → Metadata → UProp

inMap ∷ (Ord key, Eq v) ⇒ key → v → Map key v → Bool

Test if a pair is present in a map.

invertMap ∷ (Ord k, Ord v) ⇒ Map k v → Map v (Set k)

Invert a map

Examples:

>>> invertMap (Map.fromList [('a', 1 :: Int), ('b', 2), ('c', 3), ('d', 1)])
fromList [(1,fromList "ad"),(2,fromList "b"),(3,fromList "c")]

invertBijection ∷ Ord v ⇒ Map k v → Map v k

Invert a map which we assert to be a bijection. If this map is not a bijection, the behaviour is not guaranteed.

Examples:

>>> invertBijection (Map.fromList [('a', 1 :: Int), ('b', 2), ('c', 3)])
fromList [(1,'a'),(2,'b'),(3,'c')]

(==>) ∷ Bool → Bool → Bool infix 1

pvCanFollow

Arguments

∷ ProtVer
Next protocol version

→ ProtVer
Previous protocol version

→ Bool

Check whether a protocol version can follow the current protocol version.

checkUpdateConstraints ∷ PParams → UProp → [UpdateConstraintViolation]

Check whether an update proposal marks a valid update

(<=?) ∷ Ord a ⇒ a → a → Maybe (a, Threshold a) infix 4

orError ∷ Maybe (a, b) → (a → b → e) → Maybe e

canUpdate ∷ PParams → UProp → Rule UPPVV ctx ()

svCanFollow ∷ Map ApName (ApVer, Slot, Metadata) → (ApName, ApVer) → Bool

data UPIEC

data PVBUMP

data UPIEND

data UPIVOTES

data APPLYVOTES

data UPIVOTE

data UPIREG

type UPIState = ((ProtVer, PParams), [(Slot, (ProtVer, PParams))], Map ApName (ApVer, Slot, Metadata), Map UpId (ProtVer, PParams), Map UpId (ApName, ApVer, Metadata), Map UpId Slot, Set (UpId, VKeyGenesis), Set (ProtVer, VKeyGenesis), Map UpId Slot)

The update interface state is shared amongst various rules, so we define it as an alias here.

type UPIEnv = (Slot, Bimap VKeyGenesis VKey, BlockCount, Word8)

The update interface environment is shared amongst various rules, so we define it as an alias here.

data UPEND

data FADS

data UPVOTE

data ADDVOTE

vCaster ∷ Lens' Vote VKey

vPropId ∷ Lens' Vote UpId

vSig ∷ Lens' Vote (Sig UpId)

mkVote ∷ VKey → UpId → Vote

findKey ∷ (v → Bool) → Map k v → Maybe (k, v)

Find the key that corresponds to the value satisfying the given predicate. In case zero or more than one key is found this function returns Nothing.

delegationMap ∷ UPIEnv → Bimap VKeyGenesis VKey

emptyUPIState ∷ UPIState

initialPParams ∷ PParams

protocolVersion ∷ UPIState → ProtVer

protocolParameters ∷ UPIState → PParams

applicationVersions ∷ UPIState → Map ApName (ApVer, Slot, Metadata)

confirmedProposals ∷ UPIState → Map UpId Slot

futureAdoptions ∷ UPIState → [(Slot, (ProtVer, PParams))]

endorsements ∷ UPIState → Set (ProtVer, VKeyGenesis)

registeredProtocolUpdateProposals ∷ UPIState → Map UpId (ProtVer, PParams)

upiEnvGen ∷ Gen UPIEnv

dmapGen ∷ Word8 → Gen (Bimap VKeyGenesis VKey)

ppsUpdateFrom ∷ PParams → Gen PParams

Generate a protocol parameter update from a given set of current protocol-parameters, ensuring the consistency of the new protocol parameters w.r.t. the current ones, according to the canUpdate predicate in the formal specification.

TODO: we can move this into a Generator's module, but first we need to disentangle the dependencies. Moving this to Ledger.Update.Generators will cause a circular dependency. I think the rules need to be moved into their own modules.

increasingProbabilityAt ∷ Gen a → (a, a) → Gen a

Generate values the given distribution in 90% of the cases, and values at the bounds of the range in 10% of the cases.

This can be used to generate enough extreme values. The exponential and linear distributions provided by hedgehog will generate a small percentage of these (0-1%).

randomUpId ∷ Gen UpId

Generate a random update proposal id, by picking a large number so that the probability of having an update proposal with that id is nearly zero.

reSign ∷ UProp → UProp

Update the signature of the update proposal.

pickHighlyEndorsedProtocolVersion

Arguments

∷ [(ProtVer, Set VKeyGenesis)]
Current set of endorsements

→ Gen (Maybe ProtVer)

Given a list of protocol versions and keys endorsing those versions, generate a protocol-version endorsement, or Nothing if the list of endorsements is empty. The version to be endorsed will be selected from those versions that have the most endorsements.

updateProposalAndVotesGen ∷ UPIEnv → UPIState → Gen (Maybe UProp, [Vote])

Generate an optional update-proposal and a list of votes, given an update environment and state.

The update proposal and votes need to be generated at the same time, since this allow us to generate update votes for update proposals issued in the same block as the votes.

protocolVersionEndorsementGen ∷ UPIEnv → UPIState → Gen ProtVer

Generate an endorsement given an update environment and state.

goblinData_UPIREG_UPREGFailure_DoesNotVerify ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_DoesNotVerify ∷ SignalGenerator UPIREG

goblinData_UPIREG_UPREGFailure_NotGenesisDelegate ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_NotGenesisDelegate ∷ SignalGenerator UPIREG

goblinData_UPIREG_UPREGFailure_UPVFailure_AVChangedInPVUpdate ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_UPVFailure_AVChangedInPVUpdate ∷ SignalGenerator UPIREG

goblinData_UPIREG_UPREGFailure_UPVFailure_PVChangedInSVUpdate ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_UPVFailure_PVChangedInSVUpdate ∷ SignalGenerator UPIREG

goblinData_UPIREG_UPREGFailure_UPVFailure_ParamsChangedInSVUpdate ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_UPVFailure_ParamsChangedInSVUpdate ∷ SignalGenerator UPIREG

goblinData_UPIREG_UPREGFailure_UPVFailure_UPPVVFailure_CannotFollowPv ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_UPVFailure_UPPVVFailure_CannotFollowPv ∷ SignalGenerator UPIREG

goblinData_UPIREG_UPREGFailure_UPVFailure_UPPVVFailure_CannotUpdatePv ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_UPVFailure_UPPVVFailure_CannotUpdatePv ∷ SignalGenerator UPIREG

goblinData_UPIREG_UPREGFailure_UPVFailure_UPSVVFailure_AlreadyProposedSv ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_UPVFailure_UPSVVFailure_AlreadyProposedSv ∷ SignalGenerator UPIREG

goblinData_UPIREG_UPREGFailure_UPVFailure_UPSVVFailure_CannotFollowSv ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_UPVFailure_UPSVVFailure_CannotFollowSv ∷ SignalGenerator UPIREG

goblinData_UPIREG_UPREGFailure_UPVFailure_UPSVVFailure_InvalidApplicationName ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_UPVFailure_UPSVVFailure_InvalidApplicationName ∷ SignalGenerator UPIREG

goblinData_UPIREG_UPREGFailure_UPVFailure_UPSVVFailure_InvalidSystemTags ∷ GoblinData Bool

goblinGen_UPIREG_UPREGFailure_UPVFailure_UPSVVFailure_InvalidSystemTags ∷ SignalGenerator UPIREG

goblinGensUPIREG ∷ [SignalGenerator UPIREG]

goblinData_UPIVOTES_ApplyVotesFailure_UpivoteFailure_UPVOTEFailure_ADDVOTEFailure_AVSigDoesNotVerify ∷ GoblinData Bool

goblinGen_UPIVOTES_ApplyVotesFailure_UpivoteFailure_UPVOTEFailure_ADDVOTEFailure_AVSigDoesNotVerify ∷ SignalGenerator UPIVOTES

goblinData_UPIVOTES_ApplyVotesFailure_UpivoteFailure_UPVOTEFailure_ADDVOTEFailure_NoUpdateProposal ∷ GoblinData Bool

goblinGen_UPIVOTES_ApplyVotesFailure_UpivoteFailure_UPVOTEFailure_ADDVOTEFailure_NoUpdateProposal ∷ SignalGenerator UPIVOTES

goblinGensUPIVOTES ∷ [SignalGenerator UPIVOTES]

tamperWithUpdateProposal ∷ UPIEnv → UPIState → UProp → Gen UProp

Tamper with the update proposal in such a way that the following predicate failures are triggered with equal probability:

UPREGFailure

UPVFailure

UPVFailure

UPPVVFailure

CannotFollowPv

CannotUpdatePv

AlreadyProposedPv

UPSVVFailure

AlreadyProposedSv

CannotFollowSv

InvalidApplicationName

InvalidSystemTags

AVChangedInPVUpdate

ParamsChangedInSVUpdate

PVChangedInSVUpdate

NotGenesisDelegate

DoesNotVerify

tamperWithVotes ∷ UPIEnv → UPIState → [Vote] → Gen [Vote]

Tamper with some of the votes provided as parameter in such a way that the following predicate failures are triggered with equal probability:

AVSigDoesNotVerify

NoUpdateProposal

tamperWithVote ∷ Vote → Gen Vote

data family PredicateFailure a ∷ Type

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

Orphan instances

Hashable a ⇒ Hashable (Set a)

Instance details

Methods

hashWithSalt ∷ Int → Set a → Int

hash ∷ Set a → Int